+++ front-end/src/Home.css
@@ -1,12 +1,15 @@
-/* Center content on the page */
 .home-container {
+    width: 100vw;
+    max-width: 100vw;
+    padding: 20px;
+    padding-top: 400px;
+    background-color: white;
+    min-height: 100vh;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: flex-start;
-    padding: 20px;
-    background-color: #fff;
-    min-height: 100vh;
+    overflow-x: hidden;
 }
 
 /* Title for each section */
@@ -21,7 +24,7 @@
 
 /* Wrapper for each slider */
 .slider-section {
-    width: 100%;
+    width: 80%;
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -30,14 +33,14 @@
 /* Scroll container */
 .scroll-container {
     position: relative;
-    width: 80%;
-    max-width: 900px;
-    overflow: hidden; /* Hide scrollbar */
+    width:100%;
+    max-width: 100vw;
+    overflow: hidden;
     white-space: nowrap;
     padding: 10px 0;
     display: flex;
     flex-direction: row;
-    user-select: none; /* Prevent accidental text selection */
+    user-select: none;
 }
 
 /* Scroll wrapper for infinite loop */
@@ -46,17 +49,17 @@
     gap: 15px;
     flex-wrap: nowrap;
     width: fit-content;
-    transition: transform 0.1s linear; /* Smooth transition for manual scroll */
+    transition: transform 0.1s linear;
 }
 
 /* Hide scrollbar while allowing manual scrolling */
 .scroll-container::-webkit-scrollbar {
-    display: none; /* Hide scrollbar for WebKit browsers */
+    display: none; 
 }
 
 .scroll-container {
-    -ms-overflow-style: none; /* Hide scrollbar for IE/Edge */
-    scrollbar-width: none; /* Hide scrollbar for Firefox */
+    -ms-overflow-style: none;
+    scrollbar-width: none; 
 }
 
 /* Album items */
@@ -71,7 +74,7 @@
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-    cursor: pointer; /* Indicate interactivity */
+    cursor: pointer; 
 }
 
 /* Album image */
