+++ front-end/src/components/Header.css
@@ -1,0 +1,165 @@
+/* Header */
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 15px 30px;
+  background-color: white;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 90px; 
+  z-index: 1050;
+  border-bottom: 1px solid black;
+  box-sizing: border-box;
+}
+
+/* Logo and Stacked Text */
+.logo-container {
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+    background-color: transparent;
+    transition: transform 0.3s ease;
+    padding-top: 20px;
+    gap: 8px;
+  }
+
+.logo-container .logo {
+    width: 75px; 
+    height: auto;
+    display: block;
+    background: transparent;
+}
+
+/* Stacked Text */
+.stacked-text {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    color: #000;
+    font-family: "Inter", sans-serif;
+}
+
+.stacked-text h1 {
+    font-size: 24px; 
+    margin: 0;
+    margin-top: -9px;
+    font-weight: 700;
+    color: black;
+}
+
+.stacked-text h2 {
+    font-size: 18px;
+    margin-top: -3px;
+    font-weight: 700;
+    color: black;
+}
+
+/* Logo hover effect */
+.logo-container:hover {
+    transform: scale(1.1);
+}
+
+/* Profile button */
+.profile {
+    width: 33px;
+    height: auto;
+    border-radius: 50%;
+    object-fit: cover;
+    cursor: pointer;
+}
+
+.profile:hover {
+    opacity: 0.8;
+    transform: scale(1.1); 
+}
+
+/* Menu button */
+.menu {
+    cursor: pointer;
+    width: 33px;
+    height: auto;
+}
+
+/* Sidebar */
+.sidebar {
+    position: fixed;
+    top: 0;
+    left: -250px;
+    width: 250px;
+    height: 100vh;
+    background-color: black;
+    color: white;
+    padding-top: 60px;
+    transition: left 0.3s ease-in-out;
+    z-index: 1100;
+}
+
+/* Ensure sidebar does not shrink content */
+.sidebar.open {
+    left: 0;
+}
+
+.close-btn {
+    position: absolute;
+    top: 10px;
+    right: 15px;
+    font-size: 24px;
+    background: none;
+    border: none;
+    color: white;
+    cursor: pointer;
+}
+
+/* Mobile Fixes */
+@media (max-width: 768px) {
+    .header {
+      height: 70px;
+      padding: 10px 20px;
+    }
+  
+    .logo {
+      width: 70px;
+      max-height: 50px;
+    }
+
+    .stacked-text h1 {
+        font-size: 20px; 
+    }
+    
+    .stacked-text h2 {
+        font-size: 12px; 
+    }
+  
+    .menu,
+    .profile {
+      width: 35px;
+    }
+}
+  
+@media (max-width: 480px) {
+    .header {
+      height: 65px;
+      padding: 8px 15px;
+    }
+  
+    .logo {
+      width: 65px;
+      max-height: 45px;
+    }
+  
+    .stacked-text h1 {
+      font-size: 18px;
+    }
+  
+    .stacked-text h2 {
+      font-size: 10px;
+    }
+  
+    .menu,
+    .profile {
+      width: 30px;
+    }
+}
