+++ front-end/src/components/Sidebar.css
@@ -1,47 +1,56 @@
+/* Sidebar Styling */
 .sidebar {
-    position: fixed;
-    top: 0;
-    left: -250px; /* Start hidden */
-    width: 250px;
-    height: 100%;
-    background-color: #222;
-    color: white;
-    transition: left 0.3s ease-in-out;
-    padding-top: 60px;
-    box-shadow: 2px 0 5px rgba(0, 0, 0, 0.5);
-    z-index: 1000;
-  }
-  
-  .sidebar.open {
-    left: 0; /* Slide in */
-  }
-  
-  .sidebar-nav {
-    display: flex;
-    flex-direction: column;
-    padding-left: 20px;
-  }
-  
-  .sidebar-nav a {
-    color: white;
-    text-decoration: none;
-    padding: 15px 0;
-    display: block;
-    font-size: 18px;
-  }
-  
-  .sidebar-nav a:hover {
-    background-color: #444;
-  }
-  
-  .close-btn {
-    position: absolute;
-    top: 10px;
-    right: 20px;
-    font-size: 30px;
-    background: none;
-    border: none;
-    color: white;
-    cursor: pointer;
-  }
-  
+  position: fixed;
+  top: 0;
+  left: -250px; 
+  width: 250px;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.9); 
+  color: white;
+  padding-top: 60px;
+  transition: left 0.3s ease-in-out;
+  z-index: 1100;
+}
+
+/* Open Sidebar */
+.sidebar.open {
+  left: 0; 
+}
+
+/* Sidebar Navigation */
+.sidebar-nav {
+  display: flex;
+  flex-direction: column;
+  padding: 0;
+  margin: 0;
+}
+
+.sidebar-nav a {
+  display: block;
+  padding: 15px;
+  color: white;
+  text-decoration: none;
+  font-size: 18px;
+  transition: background 0.2s;
+}
+
+.sidebar-nav a:hover {
+  background: rgba(255, 255, 255, 0.2);
+}
+
+/* Close button */
+.close-btn {
+  position: absolute;
+  top: 10px;
+  right: 15px;
+  font-size: 24px;
+  background: none;
+  border: none;
+  color: white;
+  cursor: pointer;
+}
+
+/* Prevent scrolling when sidebar is open */
+body.sidebar-open {
+  overflow: hidden;
+}
